requirements.txt
opencv-python~=4.8.1.78
numpy~=1.26.4
keras~=2.13.1
pillow~=11.1.0
tensorflow~=2.13.0
pyvista~=0.46.3
matplotlib~=3.7.2
qutip~=5.2.1
scipy~=1.11.4
scikit-image~=0.25.2
albumentations~=2.0.8
scikit-learn~=1.4.2
pandas~=2.0.3
segmentation_models_pytorch~=0.5.0
torch~=2.5.1+cu121